Fixes for packaging
Added console_scripts entry to setup.py, this required a slight
refactoring of __main__.py.
Also adjusted the packages field in setup.py so it includes not only
the tellsticknet package but also tellsticknet.protocols.

Now the package can be installed with pip and the tellsticknet command
will be available on the PATH.